Add prototype of item submission while a program is running
This should be short enough to include in random long-running programs,
and shouldn't force those programs too much in a direction.

Currently, this is somewhat biased towards Mojolicious, but in principle
this approach should also work well with IO::Async.

This should not turn into a full fledged job server like Minion, so
I want to avoid persistence and cancelling. Even feedback for running
jobs is not yet mirrored to the submitting client.